# Add more runtime checks to ComparerUtilities