--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/ComparerUtilities.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/ComparerUtilities.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics;
 using Microsoft.Extensions.Internal;
 
@@ -25,6 +26,12 @@ internal static class ComparerUtilities
             return false;
         }
 
+        // PERF: Often, the IReadOnlyLists are really just arrays. For those cases, take a faster path.
+        if (first is T[] firstArray && second is T[] secondArray)
+        {
+            return AreArrayContentsEqual(firstArray, secondArray, comparer);
+        }
+
         if (first.Count != second.Count)
         {
             return false;
@@ -41,15 +48,48 @@ internal static class ComparerUtilities
         }
 
         return true;
+
+        static bool AreArrayContentsEqual(T[] first, T[] second, IEqualityComparer<T>? comparer)
+        {
+            var length = first.Length;
+
+            if (length != second.Length)
+            {
+                return false;
+            }
+
+            comparer ??= EqualityComparer<T>.Default;
+
+            for (var i = 0; i < length; i++)
+            {
+                if (!comparer.Equals(first[i], second[i]))
+                {
+                    return false;
+                }
+            }
+
+            return true;
+        }
     }
 
     public static void AddToHash<T>(ref HashCodeCombiner hash, IReadOnlyList<T> list, IEqualityComparer<T>? comparer)
     {
         comparer ??= EqualityComparer<T>.Default;
 
-        for (var i = 0; i < list.Count; i++)
+        // PERF: Often, IReadOnlyLists is array. If that's the case, take a faster path.
+        if (list is T[] array)
         {
-            hash.Add(list[i], comparer);
+            foreach (var item in array)
+            {
+                hash.Add(item, comparer);
+            }
+        }
+        else
+        {
+            for (var i = 0; i < list.Count; i++)
+            {
+                hash.Add(list[i], comparer);
+            }
         }
     }
 
@@ -78,39 +118,60 @@ internal static class ComparerUtilities
         keyComparer ??= EqualityComparer<TKey>.Default;
         valueComparer ??= EqualityComparer<TValue>.Default;
 
-        // 🐇 Avoid enumerator allocations for Dictionary<TKey, TValue>
-        if (first is Dictionary<TKey, TValue> firstDictionary
-            && second is Dictionary<TKey, TValue> secondDictionary)
+        switch ((first, second))
         {
-            using var firstEnumerator = firstDictionary.GetEnumerator();
-            using var secondEnumerator = secondDictionary.GetEnumerator();
-            while (firstEnumerator.MoveNext() && secondEnumerator.MoveNext())
-            {
-                if (!keyComparer.Equals(firstEnumerator.Current.Key, secondEnumerator.Current.Key)
-                    || !valueComparer.Equals(firstEnumerator.Current.Value, secondEnumerator.Current.Value))
+            case (Dictionary<TKey, TValue> firstDictionary, Dictionary<TKey, TValue> secondDictionary):
                 {
-                    return false;
-                }
-            }
+                    // 🐇 Avoid enumerator allocations for Dictionary<TKey, TValue>
+                    using var firstEnumerator = firstDictionary.GetEnumerator();
+                    using var secondEnumerator = secondDictionary.GetEnumerator();
+                    while (firstEnumerator.MoveNext() && secondEnumerator.MoveNext())
+                    {
+                        if (!keyComparer.Equals(firstEnumerator.Current.Key, secondEnumerator.Current.Key)
+                            || !valueComparer.Equals(firstEnumerator.Current.Value, secondEnumerator.Current.Value))
+                        {
+                            return false;
+                        }
+                    }
 
-            Debug.Assert(!firstEnumerator.MoveNext() && !secondEnumerator.MoveNext(), "We already know the collections have same count.");
-            return true;
-        }
-        else
-        {
-            using var firstEnumerator = first.GetEnumerator();
-            using var secondEnumerator = second.GetEnumerator();
-            while (firstEnumerator.MoveNext() && secondEnumerator.MoveNext())
-            {
-                if (!keyComparer.Equals(firstEnumerator.Current.Key, secondEnumerator.Current.Key)
-                    || !valueComparer.Equals(firstEnumerator.Current.Value, secondEnumerator.Current.Value))
+                    Debug.Assert(!firstEnumerator.MoveNext() && !secondEnumerator.MoveNext(), "We already know the collections have same count.");
+                    return true;
+                }
+
+            case (ImmutableDictionary<TKey, TValue> firstDictionary, ImmutableDictionary<TKey, TValue> secondDictionary):
                 {
-                    return false;
-                }
-            }
+                    // 🐇 Avoid enumerator allocations for ImmutableDictionary<TKey, TValue>
+                    using var firstEnumerator = firstDictionary.GetEnumerator();
+                    using var secondEnumerator = secondDictionary.GetEnumerator();
+                    while (firstEnumerator.MoveNext() && secondEnumerator.MoveNext())
+                    {
+                        if (!keyComparer.Equals(firstEnumerator.Current.Key, secondEnumerator.Current.Key)
+                            || !valueComparer.Equals(firstEnumerator.Current.Value, secondEnumerator.Current.Value))
+                        {
+                            return false;
+                        }
+                    }
 
-            Debug.Assert(!firstEnumerator.MoveNext() && !secondEnumerator.MoveNext(), "We already know the collections have same count.");
-            return true;
+                    Debug.Assert(!firstEnumerator.MoveNext() && !secondEnumerator.MoveNext(), "We already know the collections have same count.");
+                    return true;
+                }
+
+            default:
+                {
+                    using var firstEnumerator = first.GetEnumerator();
+                    using var secondEnumerator = second.GetEnumerator();
+                    while (firstEnumerator.MoveNext() && secondEnumerator.MoveNext())
+                    {
+                        if (!keyComparer.Equals(firstEnumerator.Current.Key, secondEnumerator.Current.Key)
+                            || !valueComparer.Equals(firstEnumerator.Current.Value, secondEnumerator.Current.Value))
+                        {
+                            return false;
+                        }
+                    }
+
+                    Debug.Assert(!firstEnumerator.MoveNext() && !secondEnumerator.MoveNext(), "We already know the collections have same count.");
+                    return true;
+                }
         }
     }
 
@@ -120,22 +181,36 @@ internal static class ComparerUtilities
         keyComparer ??= EqualityComparer<TKey>.Default;
         valueComparer ??= EqualityComparer<TValue>.Default;
 
-        // 🐇 Avoid enumerator allocations for Dictionary<TKey, TValue>
-        if (dictionary is Dictionary<TKey, TValue> typedDictionary)
+        switch (dictionary)
         {
-            foreach (var kvp in typedDictionary)
-            {
-                hash.Add(kvp.Key, keyComparer);
-                hash.Add(kvp.Value, valueComparer);
-            }
-        }
-        else
-        {
-            foreach (var kvp in dictionary)
-            {
-                hash.Add(kvp.Key, keyComparer);
-                hash.Add(kvp.Value, valueComparer);
-            }
+            case Dictionary<TKey, TValue> typedDictionary:
+                // 🐇 Avoid enumerator allocations for Dictionary<TKey, TValue>
+                foreach (var (key, value) in typedDictionary)
+                {
+                    hash.Add(key, keyComparer);
+                    hash.Add(value, valueComparer);
+                }
+
+                break;
+
+            case ImmutableDictionary<TKey, TValue> typedDictionary:
+                // 🐇 Avoid enumerator allocations for ImmutableDictionary<TKey, TValue>
+                foreach (var (key, value) in typedDictionary)
+                {
+                    hash.Add(key, keyComparer);
+                    hash.Add(value, valueComparer);
+                }
+
+                break;
+
+            default:
+                foreach (var (key, value) in dictionary)
+                {
+                    hash.Add(key, keyComparer);
+                    hash.Add(value, valueComparer);
+                }
+
+                break;
         }
     }
 }